ia64: unwind: properly guard UNW_DPRINT calls with UNW_DEBUG
The original implementation relies on UNW_DPRINT being defined to null to
skip debug output (when UNW_DEBUG is not defined), which resulted in two
instances of empty else branches. In those two instances, when the else
brace is not surrounded by braces, it then triggered `-Werror=empty-body'.

Adding the braces, however, would not make the code any more sane, as
UNW_DEBUG is never defined in the build system. To rectify this, introduce
a Kconfig option named IA64_UNW_DEBUG, which, when enabled passes
-DUNW_DEBUG=0 to make this behaviour actually controllable (and consistent
with the original semantics).

Signed-off-by: Cyan <cyan@cyano.uk>
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>